requirements.txt
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
dnspython==2.0.0
idna==2.10
lxml==4.5.2
nslookup==1.1.0
pygeoip==0.3.2
requests==2.24.0
soupsieve==2.0.1
urllib3==1.25.10
whois==0.9.7
